Refactor app structure, add routing with react-router-dom
