+++ package.json
@@ -43,21 +43,22 @@
     "@goperigon/perigon-ts": "^1.0.5",
     "@marsidev/react-turnstile": "^1.1.0",
     "@modelcontextprotocol/sdk": "^1.13.1",
-    "agents": "^0.0.97",
-    "@vitejs/plugin-react": "^4.6.0",
-    "ai": "^4.3.16",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-markdown": "^10.1.0",
-    "zod": "^3.25.67",
-    "zod-to-json-schema": "^3.24.6",
     "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-collapsible": "^1.1.11",
     "@radix-ui/react-select": "^2.2.5",
     "@radix-ui/react-slot": "^1.2.3",
+    "@vitejs/plugin-react": "^4.6.0",
+    "agents": "^0.0.97",
+    "ai": "^4.3.16",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.525.0",
-    "tailwind-merge": "^3.3.1"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-markdown": "^10.1.0",
+    "react-router-dom": "^7.6.3",
+    "tailwind-merge": "^3.3.1",
+    "zod": "^3.25.67",
+    "zod-to-json-schema": "^3.24.6"
   }
 }
