+++ src/pages/ChatPage.tsx
@@ -1,0 +1,5 @@
+import ChatPlayground from "@/components/chat-playground";
+
+export default function ChatPage() {
+  return <ChatPlayground />;
+}
